update for challenge
#Election Audit Results

##Overview and Purpose of Project

The Election Result data provides us with election outcomes based on each voter’s ID. The raw data shows who each person voted for, and what county they were in. It is up to us to summarize this data, based on candidate and county votes, to determine who won the election. 

###Analysis of the Counties

•	There were a total of  363,711 votes cast 
•	Denver county received 83% of the votes, making it the highest turnout within the 3 counties. 306 thousand people voted in Denver
•	Jefferson received 11% of the votes, with almost 39 thousand voters.
•	Arapahoe county received 6.7% of the votes, with almost 24 thousand voters.

### Analysis of the Candidates

•	Dianna DeGette is the winner of this election, receiving 272,892 votes.
•	Dianna Degette received 73.8% of the votes so it is safe to say she won by a landslide.
•	Charles Casper Stockham received 23% of the votes, totaling 85k.
•	Raymon Anthony Doane received 3.1% of the votes, totaling 11.6k

##Election Audit Summary

In summary, the data was able to show us the voter turnout by county and the candidate winning results. This code can be reused again to evaluate election results in any election. Perhaps, in the next election we want to learn more information on why a candidate voted for who they did. We can look at factors like race, employment status, and age. 

We could use the code to create a race dictionary. The race dictionary could tell us how many people of that specific race came out to vote. A for loop would count the number of races and we would be able to see the percentage turnout by race. This could show us if the election was swayed based on the race of the voters. 

We could also use the code to iterate through employment status. To code this, we would create a for loop and count the number of employed and unemployed people. We could see if the employment status of the voters had any relation to the outcome of the candidates.